requirements.txt
numpy
streamlit
opencv-python
google-generativeai
SpeechRecognition
pyttsx3